\name{ExampleData}
\alias{prostateTMrisk}
\alias{diabetes1}
\alias{diabetes2}
\docType{data}

\title{The example data in the environment variable of the system}
\description{
  The example data in the environment variable of the system.
}
\details{
The example data includes the variable \code{prostateTMrisk}, \code{diabetes1} and \code{diabetes2}. The \code{prostateTMrisk} is the differential metabolites from the study of Sreekumar et al., the \code{diabetes1} from text mining and the \code{diabetes2} is the differential metabolite set from the study of Suhre et al.. These three example datasets can be obtained by the function \code{\link{GetExampleData}}.

}

\references{
Sreekumar, A., Poisson, L.M., Rajendiran, T.M., Khan, A.P., Cao, Q., Yu, J., Laxman, B., Mehra, R., Lonigro, R.J., Li, Y. et al. (2009) Metabolomic profiles delineate potential role for sarcosine in prostate cancer progression. Nature, 457, 910-914.

Suhre, K., Meisinger, C., Doring, A., Altmaier, E., Belcredi, P., Gieger, C., Chang, D., Milburn, M.V., Gall, W.E., Weinberger, K.M. et al. (2010) Metabolic footprint of diabetes: a multiplatform metabolomics study in an epidemiological setting. PLoS ONE, 5, e13953.
}
\format{An environment variable}
\author{Yanjun Xu <tonghua605@163.com>, Chunquan Li <lcqbio@aliyun.com.cn> and Xia Li <lixia@hrbmu.edu.cn>}
\keyword{file}