\name{MPINet-internal}
\alias{performpcls}
\alias{fdr.est}
\alias{initializeMPINet}
\alias{Getenvir}
\title{MPINet internal functions}
\description{
  These functions internally called by other functions in the package \code{MPINet} and not meant to be called by the user.
}

\usage{
    performpcls(x, y, nKnots = 6)
    fdr.est(p)
    initializeMPINet()
    Getenvir(envirdata)
}

\author{Yanjun Xu <tonghua605@163.com>, Chunquan Li <lcqbio@aliyun.com.cn> and Xia Li <lixia@hrbmu.edu.cn>}

\keyword{internal}
